requirements.txt
# Core Backend

fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0

# LLM / AI

openai>=1.40.0
langchain>=0.3.0
langchain-openai>=0.2.0

# Vector Database

faiss-cpu>=1.8.0
numpy>=1.26.0

# Embeddings

sentence-transformers>=3.0.0

# HuggingFace NLP

transformers>=4.41.0

# Speech-to-Text

openai-whisper
torch>=2.1.0
ffmpeg-python

# Web UI

streamlit>=1.39.0

# Utilities

httpx>=0.27.0
python-dotenv>=1.0.1
pandas>=2.2.0

# NLP / Topic Extraction

keybert>=0.8.5
scikit-learn>=1.4.0
spacy>=3.7.0
